Added multiple functions and cleaned up code
Run database.exe and type `help` to be shown the list of available commands and what they do.
Made the C++ code cleaner and more readable, added comments, and a new 'headers.h` file to organize things.
From here, there are multiple things to futher accomplish:
1: Add more variables to the `Record` class. This would likely include a list of publications by the professor which can be accessed on the website by a dropdown list for the viewer to browse (and thus for the viewer to determine whether they are interested in the prof's research).
2: Begin construction of a server (either physical or remote) that will constantly run `database.exe` to A) recieve scraped prof data from the internet and B) auto-save and auto-compile to update the website and backup the data. I have a friend who suggested Google Collab, but I was also thinking of commissioning him to construct a physical server.
3: Finally start styling the website. I guess its time for me to learn HTML/CSS (or get another person on board for that).
A few other notes:
For point (2), I was able to transfer and compile the database onto a linux OS (Ubuntu WSL) and it ran very well there. This would be the likely course of action should we start having a server running `database.exe` 24/7.
When entering data via the `add` or `edit` functions, PLEASE avoid using the `|` symbol (this is the delimiter for the fields), and if you need to type an apostraphe ', put a backslash before it \'. This is to avoid errors in the js compiler.